+++ src/components/SmartReview/Acknowledgements.js
@@ -1,0 +1,20 @@
+import UserAvatar from 'components/UserAvatar/UserAvatar';
+import React from 'react';
+import { useSelector } from 'react-redux';
+
+const Acknowledgements = () => {
+    const contributors = useSelector(state => state.smartReview.contributors);
+
+    return (
+        <div className="d-flex mb-4">
+            {contributors &&
+                contributors.map(contributor => (
+                    <div className="mr-1" key={contributor}>
+                        <UserAvatar userId={contributor} size={40} />
+                    </div>
+                ))}
+        </div>
+    );
+};
+
+export default Acknowledgements;